Refactor reporter to interfaces (#345)
The reporter has been refactored to be an interface with various
implementations. Based on the idea from Michael here:
https://github.com/google/osv-scanner/pull/341#discussion_r1164891532

Want to get some feedback on whether this is worth changing or not.

I'm also not sure if we should have the different reporter types public
here, or make them private and have the users of the library provide
their own implementations of Reporter.