Updated to reflect cvss scores being added to output table. (#419)
Addresses #418

I added CVSS to the output page. I calculated the scores using
[this](https://www.first.org/cvss/calculator/3.1#CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:N/I:N/A:H).
I think it is doing the same thing as we are, but I have a note to
confirm with our tool once it is released.

I added a section on the usage page that lists everything that is
outputted, so I can say that CVSS is calculated using the
[severity[].score](https://ossf.github.io/osv-schema/#severity-field)
field. I am open to other ideas for how to present that information
though.

[Preview available here](https://hayleycd.github.io/osv-scanner/)

---------

Signed-off-by: Hayley Denbraver <denbraver@google.com>
Co-authored-by: Rex P <106129829+another-rex@users.noreply.github.com>