Add go stdlib version support (#484)
Fixes #453,

As discussed in #453, it does require go to be installed on the system
that osv-scanner is being ran on. The go installation can be at a custom
location as long as $GOROOT environment var is specified.

Also exclude the depguard check, which has been added in a newer version
of golangci (which we haven't updated to, but will cause linting issues
when we eventually do).