Add osv output lockfile + refactor (#505)
Introduces the osv-scanner results lockfile format (feel free to suggest
a better name for this),

This also refactors the models folder slightly to move some of the
helper functions out to a separate internal crate. This stops the models
package from pulling in the lockfile package, causing cyclic imports.

I also added another field to `PackageDetails`, `Source`, a currently
optional string to specify the location the dependency is actually
specified. The only times it will be different from the lockfile
location currently is:

- Python requirements.txt `-r`
- This new osv-scanner format, it stores the lockfile path parsed in the
output results.