Revert breaking change in `osv.go` (#514)
Changing the type of Package.Ecosystem would be a breaking change, so
revert that back to string. Doesn't affect too much so looks fine to me,
let's add this to #442 for V2.

Also changed `Response.Vulns` back to type `[]models.Vulnerability`. I
don't think this changes anything since go coerces the types to each
other when setting the value.

And made the not supported error a bit more specific