✨ Add --exclude-packages Flag to osv-scanner CLI
This PR adds a --exclude-packages flag to the OSV-Scanner CLI.
This flag allows users to specify a list of regex patterns to which
matches of collected package names will be matched against.

If the packages are found to match the regex patterns, they will be
excluded from the query to the OSV database.

Exclusions work for both lockfile and SBOM package definitions.