license checking experimental feature (#501)
This experimental feature calls the deps.dev API for licenses on each package. If an allowlist is specified, it reports on packages with violating licenses.

An --all-packages flag is also added, which causes all packages to be included in the json even if it doesn't have any issues (vulns or license violations).