Fix filter ids (#647)
Fixes #634 

The actual change is just adding an `Aliases` field to the Group output,
that combines all the IDs and aliases together. A lot of fixtures had to
be updated though.

Added an additional test for this in `main_test`, and also modified a
test in `osvscanner_internal_tests.go`

Also added `omitempty` tag to `PackageInfo.commit` which it should have
contained in the first place.