Add option to not fail on vulnerability being found for github action (#732)
Add an option for the github action reporter to not return a failing
code when vulnerabilities are found. This can be used to switch the
osv-scanner action into a more "informative" mode rather than forcing
strictly 0 vulnerabilities.

The success/failure of a workflow can now indicate whether the workflow
itself completely successfully and uploaded scan results, rather than
showing the contents of the results.

A followup PR will be created adding the option to the workflows once
this is merged.