Assume latest patch version if version does not exist (#740)
While the change made in #704 is the ideal solution, it only works for
people using go 1.21 or later. Adding a patch version to go.mod will
cause go1.20 or earlier to fail to build. Since 1.20 is still fully
supported and is still the default version for many workflows, this
causes too many breakages for users to be acceptable.

There's a few alternative fixes:  

- Assume latest patch version (This PR)
- This means osv-scanner is no longer checking whether a project is
updating to the latest patch version to resolve vulnerabilities, but
will still catch users using an out of support go minor version that has
no patch available.
- Remove stdlib vulnerability check entirely / move stdlib vuln check
behind a experimental flag
- Add a way to hint to osv-scanner the go version (e.g. .go-version
file, or a comment in the go.mod file specifically for osv-scanner to
check).

I am leaning towards assuming the latest patch as a fix in the mean time
until go1.22 releases (at which point 1.20 will no longer be supported),
and less people will be using 1.20, and perhaps reintroducing the patch
version check.

---

This implementation also means when a stdlib vulnerability is found, the
version that is printed will be 1.xx.99, which is not quite accurate.