Build action image when releasing (#747)
Automatically build the github action image as part of the goreleaser
pipeline. The image name will be github.com/google/osv-scanner-action ,
but the image/package will be associated with the osv-scanner
repository.

This might fail if there is a permission issue when building/uploading
the image in the release, will explore during release.