Update goreleaser to fix release pipeline (#751)
Goreleaser only puts the built binaries in the context.
(https://goreleaser.com/errors/docker-build/) This causes the custom
github action docker build to fail.

This PR adds osv-reporter as something goreleaser builds, and correctly
copies in the bash script. Tested to build locally.

I also updated the goreleaser.yml file to exclude the osv-reporter in
the output, but I don't know if that will be successful as there's no
way to test it.