Guided Remediation: Compute in-place updates (#789)
Adding in-place update computation, for finding 'drop-in' fixes of vulns
in indirect dependencies.

I did end up rewriting a lot of this, so I'd appreciate a more thorough
review of this PR.

Something to note for the future: unlike what happens internally, the
returned patches can be incompatible with each other (i.e. could
possibly suggest two different versions for the same package). I'll need
to make sure the caller handles this possibility.