ci: remove unneeded `setup-go` step and pin `actions/download-artifact` (#786)
I was having a look at the openssf scorecard score as I've not really
dug into that before, and saw it was flagging some unpinned actions.

Most of them were about the `osv-scanner` action itself which I don't
think can be pinned since the point is to use the latest version for
self testing right? but we can pin `actions/download-artifact` and
`actions/setup-go` _though_ I'm pretty sure that isn't needed anymore
because it was a workaround for
https://github.com/github/codeql/issues/13992 and we were just waiting
on a new version? either way we'll find out.