Replace stereoscope with using go-containerregistry directly (#836)
Replace stereoscope with using the lower level go-containerregistry
library directly.

Stereoscope brings in a large list of dependencies to support functions
that is not necessary for osv-scanner. This also gives us more low level
control of the extraction and scanning.

It's possible to switch to a new fs.FS lockfile interface without too
much work.