Find and save the distro version when extracting from debian and alpine (#854)
Previously, we ignored the distro version when querying for
vulnerabilities. This adds in parsing for distro version.

To enable this for alpine, which doesn't encode their release version in
their `apk-installed` file, I also implemented `Open()` function for
image `DepFile`, and for alpine, open the `/etc/alpine-version` file to
read the alpine version.

This hid the previous vulnerabilities that were showing up for alpine
3.10, since alpine 3.10 is no longer supported, the advisories weren't
being published for those packages, even though those vulnerabilities
probably do apply.

---

It is quite difficult to find packages that are actually vulnerable and
haven't been updated in alpine, and older packages are removed from the
repository, so we can't really test the positive case where a
vulnerability is found.