Put API/networking errors on another error code (#857)
Separate out API errors into its own error code. The intention is to
avoid lumping everything in error 127, since it is reserved for general
errors not actionable by the user.

Errors related to the API should be separate as they are actionable by
the user (e.g. the CI it's running on has no access to the internet), or
our API is timing out, so they should try again...etc.

Another motivation is that currently because API errors are mixed in
with general errors, it's not clear to the user what caused the
osv-scanner GitHub action to fail, since we specifically ignore general
errors and set the exit code to 0 in the GitHub action, as this could be
caused by failing to parse a specific lockfile. This leads to situations
like:
https://github.com/google/osv.dev/actions/runs/8243615203/job/22544495994?pr=2048
, where it's unclear what the actual error is (it is hiding under `Run
Scanner For Existing Code` step, which is a check).

This PR also clarifies the error wording to make it clear which query
caused the error.

TODO: (In followup PRs)
- Guided Remediation
- Automated updates