GR: Add tests for npmrc & npm registry api (#879)
Add some tests for `npmrc` parsing / registry request building and
querying of the npm registries.

Currently, the `npmrc` parser can use a few environment variables, which
makes it a bit difficult to test properly. I may change it in another PR
to make it take them in as an argument so I can sandbox the tests
better.