GR: Add test universe generation script and tests for patch generation (#967)
Made a go script that creates a universe file for the
`MockResolutionClient` (#909) from the current real deps.dev & OSV data
used during in-place / relock / relax computation.

This has let me add some real-world in-place & relax patch generation
test cases to test for regression without it breaking constantly due to
new versions. I've only added one universe because the files are pretty
large.

The in-place/relax patch generation tests are basically end-to-end tests
- it's not really written in a way that makes testing only the patch
computation functionality possible. (This might be a code smell, but I
don't think it's practical to structure the code in such a way).