Fix some Maven manifest & resolver issues (#1008)
Some fixes for a few Maven resolver issues I've come across:
1. Requirement origins weren't being tracked correctly if a package key
was set by a property. Fixed by checking dependency keys before and
after interpolation, and updating the map if they changed. (I've
modified one of the tests to check for this case)
2. To work around the resolver not resolving test or optional
dependencies, made it so the pom.xml parser removes the test scope and
optional flags.
3. `resolve.PackageKey` was not sufficient to uniquely key the
requirements for the `Groups` map in the `Manifest`. Made a new
`RequirementKey` type with ecosystem-specific information for both npm
and maven to solve this.