✨  Adding CycloneDX 1.4 and 1.5 reporter (#1014)
## About this PR

This PR aims to add the CycloneDX 1.4 and 1.5 format. The idea behind it
is to be able to export everything in a commonly used format to later
reuse the result to enrich it. (ex: re-running a vulnerability matching
process on the packages to find emerging vulnerabilities).

This PR adds 2 versions of CycloneDX even though at the moment they
report the exact same thing.

In future PR, we intend to extract where packages has been found during
the extraction process (file, line, column), which is an information
only supported by CycloneDX 1.5.

We need a CycloneDX 1.4 format for backward compatibility reasons.

## Additional informations

You may want to review this PR commit by commit, as it should be easier
to read