feat(guided remediation): remediate unresolved dependency management vulns (#1235)
Adds functionality to allow guided remediation to fix vulns in
`dependencyManagement` dependencies that do not appear in the resolved
dependency graph of the POM - useful for 'remediating' POMs without any
actual dependencies.

I've accomplished this by checking if each of the original management
dependencies (*excluding* those inherited from parents) appear in the
graph after the initial resolution. If they're missing, I add them to
the graph as direct dependencies (not resolving their transitive
dependencies).

This behaviour is disabled by default, and I've added a
`--maven-fix-management` flag to enable it. I was going to try combine
this and `--ignore-dev` into a `--groups` flag but it seemed like it
would be a bit too complicated.