feat(guided remediation): support offline database in fix subcommand (#1306)
#1295
Support the `--experimental-offline` and
`--experimental-download-offline-databases` flags in `osv-scanner fix`.

This is pretty inefficient since `local.MakeRequest()` reads and parses
the zip file every time it's called. I didn't want to make too many
changes to make it more efficient since we'll probably end up
refactoring it in v2.

Also fixed the links to headings on the offline docs page.