ci: add versions fixture generator for Red Hat (#1356)
This adds a `semantic` fixture generator for the Red Hat ecosystem - as
this ends up being just under 2mb, I've not committed the file like the
others, instead making the test suite only use it if it's actually
present.

It works by leveraging [`rpm`s built-in Lua
interpreter](https://rpm-software-management.github.io/rpm/manual/lua.html),
which includes exposing the `vercmp` function meaning we can have `rpm`
tell us what the result of comparing two versions are so long as there's
a version of that available.

In terms of actually getting a version of `rpm`, that turns out to be a
lot harder with Docker I assume just because CentOS & co is a less
open-source-y distro vs the others; there are images, but they're all
generally very big. In an interesting twist though it turns out Debian
actually ships a version of `rpm` for building and the like - it has had
it's package management abilities neutered, but that's not something we
need so a simple `apt install rpm` gives us the dependency we need.