fix: Fix VCS extractor and suppresses vcs errors  (#1458)
Currently the extractor is opening the .git directory, when the
extractor should be opening the parent directory.

Also changed so on empty repositories, (or empty submodules) it will not
return an error.
This is following the osv-scalibr convention where if an extractor could
fail to extract on valid files, and the error is sort of expected to
happen some of the time, it should not return an error.

Ideally we might want to have this type of error return a special error
type.

FYI  @spencerschrock This fixes the error discussed in #1438