feat: Vendored extractor (#1451)
This PR adds the vendored extractor. The contents under testdata is zlib
vendored in with a couple of files removed.

If there is a better way to do this without clogging up the number of
files git istracking, please let me know.

Importantly, this is not yet enabled in the actual extraction, as there
are bugs in the walker which will be resolved in a separate PR.