fix: walker now properly uses Root and passes non absolute paths (#1452)
fs.FS is supposed to use only non absolute paths, and be rooted at the
real path Root. This change does this.

As a consequence of this, some error messages will not be rooted at
Root, so also find and replace without the initial leading /. I wonder
how this will work on windows though.

Follow up TODO is to update the pomxml extractor to use the virtual
filesystem, rather than using the real filesystem with the Root
shortcut.