fix(maven): do not track registries when fetching requirements (#1466)
https://github.com/google/osv-scanner/issues/1447

When fetching requirements of a Maven dependency, we should not keep
track of the registries defined in pom.xml of dependencies. Doing that
will add a lot irrelevant registries to the client and sends unnecessary
requests, and finally slows down the resolution.

This PR also disables `PreFetch` in Maven transitive scanning.
`PreFetch` calls `MatchVersions` which sends requests to
`maven-metadata.xml` for native Maven client. `maven-metadata.xml` is
not necessarily needed for all dependencies. Disabling `PreFetch` will
save us from making requests to these files. However, calling `PreFetch`
should be fine for deps.dev client, and this can be a TODO in the
future.

With this fix, resolving
[pom.xml](https://github.com/google/osv-scanner/blob/main/internal/remediation/fixtures/zeppelin-server/pom.xml)
is now shortened to 10-20 seconds.