feat(guided remediation): non-interactive JSON output format (#1437)
Closes #1213
Adds `--format` flag to `osv-scanner fix --non-interactive` to choose
between old `text` and new `json` format, which provides a more
machine-readable description of the changes made by guided remediation.
The `text` format should remain unchanged.

Also, renamed the `relock` strategy to `relax` (though maintained
backwards-compatibility) to be a bit less ambiguous.

There are a few things that'll need to change with osv-scanner v2. In
particular, the way the reporter interface is being used. I'll need to
revisit this when those are refactored.