refactor: imodels PackageInfo refactor to use functions (#1482)
This PR refactors PackageInfo to be accessed via methods, which will
perform necessary transformations on the underlying inventory as needed.

This also adds the ecosystemmock extractor which allows you set the
return value of the Ecosystem() function.

TODO: Add tests for imodels. I will add tests in a followup PR once we
are confident we are sticking with this approach and it works with the
other refactors.