feat: Experimental code to start java class reachability. (#1480)
This adds the beginnings of a Java reachability analyser that
recursively parses class constant pools
(https://docs.oracle.com/javase/specs/jvms/se22/html/jvms-4.html#jvms-4.7)
for references to other classes.

The intention is to see if this can be used to exclude transitive
dependencies
from vulnerability scanning completely, if they can be proven to be
unreachable.

There's still a lot of TODOs remaining (recorded in the Go source as
comments). This tool currently expects
all dependency class files to be available (passed via `--classpath`).