fix(guided remediation): npm workspaces devDeps in in-place strategy (#1512)
`devDependencies` can be declared in a workspace, which is not a direct
dependency (`root` -> `workspace` -> `dev`). Made the IsDev check that
`in-place` uses also check these so that `--ignore-dev` works properly
for `in-place` with npm workspace.
(`--ignore-dev` still doesn't behave correctly for `relax` with
workspaces, but relax doesn't really work with workspaces anyway)

Also fixed a potential panic with in-place & private registries that
don't behave as expected w.r.t. `optionalDependencies` being in regular
`dependencies`