fix(guided remediation): remove `--relock-cmd` flag (#1517)
`--relock-cmd` let you override the `npm install` command with an
arbitrary command to regenerate the `package-lock.json` after doing a
relax.
We probably don't want to have that option in osv-scalibr when we
migrate this there, so I've removed it now before the v2 release of
osv-scanner.

If someone actually needs this, they can run the command themselves
outside of osv-scanner.